클래스 패키지 이동 및 관계 설정
